do study nacos —— nacos project CRUD implement

+++ 04-jt-admin/jt-modules/jt-system/src/main/java/com/abo/jt/system/dao/SysNoticeDao.java
@@ -46,5 +46,5 @@ public interface SysNoticeDao {
      * @param id
      * @return
      */
-    SysNotice selectById(Integer id);
+    SysNotice selectById(Long id);
 }
